requirements.txt
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coverage==4.5.2
idna==2.8
nose==1.3.7
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
